+++ packages/orq-rc/RELEASES.md
@@ -7919,3 +7919,13 @@ Based on:
 - [typescript v4.7.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.35 - packages/orq-rc
+
+## 2026-04-01 03:06:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.36 - packages/orq-rc